+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 
 ### Added
 
+- ✨(dimail) synchronize mailboxes from dimail to our db #453
 - ✨(ci) add security scan #429
 - ✨(teams) register contacts on admin views